requirements.txt
flask
requests
psycopg2-binary
pinecone-client
pgvector
python-dotenv
